Major cleanup of attributes logic
