+++ cli/src/builders/make/folderSettings.ts
@@ -1,0 +1,35 @@
+import * as path from "path";
+import { getFiles } from "../../utils";
+import { readFileSync } from "fs";
+import { warningOut } from "../../cli";
+
+export interface FolderOptions {
+	version?: "0.0.1",
+	build?: {
+		// objlib?: string, We don't support objlib
+		tgtCcsid?: string
+	}
+}
+
+/**
+ * Fetch the folder specific settings for a project
+ */
+export function getFolderOptions(cwd: string) {
+  // Then fetch the directory specific settings
+  const ibmiFiles = getFiles(cwd, `**/.ibmi.json`);
+
+  let folderSettings: { [key: string]: FolderOptions } = {};
+
+  for (const ibmiFile of ibmiFiles) {
+    const relative = path.relative(cwd, ibmiFile);
+    const folder = path.dirname(relative);
+
+    try {
+      folderSettings[folder] = JSON.parse(readFileSync(ibmiFile, { encoding: `utf-8` }));
+    } catch (e) {
+      warningOut(`make: Failed to read ${relative}.`);
+    }
+  }
+
+  return folderSettings;
+}
